Another commit.
Floppy, HDD and CD/DVD driver for Booter will be made soon, now loading kernel
from initrd (yeah, it will hit it's limit at some time). Paging initialization code written,
however it crashes in random places. I'll do some debugging next week.

+++ loader/stage3/booter.h
@@ -27,7 +27,7 @@ namespace Booter
     // print panic message
     extern void Panic(char *);
     
-    extern int GetTimestamp();
+    extern unsigned long long int GetTimestamp();
     
     extern InitRD::InitRD * initrd;
 }
